♻️ refactor: summary types
