+++ src/components/@shared/Navbar.tsx
@@ -12,7 +12,6 @@ import CallMadeIcon from "@mui/icons-material/CallMade";
 import { Button } from ".";
 import useUser from "src/hooks/useUser";
 import { BLACK, WHITE } from "src/constants/colors";
-import Image from "next/image";
 
 function NavBar() {
   const { isLoggedIn, logout } = useUser();
