+++ src/components/ProjectList/ProjectCard.tsx
@@ -1,26 +1,21 @@
 import { useRouter } from "next/router";
 import { MouseEvent } from "react";
-import { useSetRecoilState } from "recoil";
 import { Box, CardContent, CardActionArea, Typography } from "@mui/material";
 import {
   ChangeHistory as ChangeHistoryIcon,
-  Close as CloseIcon,
   GitHub as GitHubIcon,
 } from "@mui/icons-material";
 
 import timeSince from "src/utils/timeSince";
 import useModal from "src/hooks/useModal";
-import { selectedProject } from "src/recoil/userDeployments";
 import { BLUE } from "src/constants/colors";
 
-import type { Project } from "./ProjectCardList";
+import type { Project } from "types/api";
 interface ProjectCardProps {
   project: Project;
 }
 
 function ProjectCard({ project }: ProjectCardProps) {
-  const { showModal } = useModal();
-
   const updatedMilliseconds = new Date(project.projectUpdatedAt).valueOf();
   const repoUpdatedSince = timeSince(updatedMilliseconds);
   const router = useRouter();
@@ -28,17 +23,6 @@ function ProjectCard({ project }: ProjectCardProps) {
   const handleCardClick = () => {
     const { space, repoName } = project;
     router.push(`/${space}/${repoName}`);
-  };
-
-  const handleCloseButtonClick = (e: MouseEvent) => {
-    e.stopPropagation();
-
-    showModal({
-      modalType: "ModalDeleteConfirm",
-      modalProps: {
-        project,
-      },
-    });
   };
 
   const handleClickUrl = (e: MouseEvent) => {
@@ -50,7 +34,6 @@ function ProjectCard({ project }: ProjectCardProps) {
       <CardContent
         sx={{ position: "relative", padding: "1.5rem", height: 180 }}
       >
-        {/* <CloseIcon sx={CloseIconStyle} onClick={handleCloseButtonClick} /> */}
         <Box
           sx={{ display: "flex", flexDirection: "row", alignItems: "center" }}
         >
@@ -96,17 +79,5 @@ function ProjectCard({ project }: ProjectCardProps) {
     </CardActionArea>
   );
 }
-
-const CloseIconStyle = {
-  position: "absolute",
-  top: "0.5rem",
-  right: "0.5rem",
-  color: "rgba(51, 51, 51, 0.8)",
-  width: "1.2rem",
-  "&:hover": {
-    color: "rgba(255, 83, 83, 0.8)",
-    backgroundColor: "white",
-  },
-};
 
 export default ProjectCard;
