+++ src/types/global.d.ts
